Split NotebookPost into content + shared components files
